youtube-summarizer/pyproject.toml
[project]
name = "youtube-summarizer"
version = "0.1.0"
description = "A YouTube video summarizer that extracts transcripts and summarizes them using GPT-4o"
authors = [
    {name = "AI Builders Bootcamp", email = "your-email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "youtube-transcript-api>=0.6.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
